update account display to detach it from the table move selection management to parent component - RedisCloudSubscriptionsPage.tsx extracted logic and configuration to a hook
