+++ redisinsight/ui/src/pages/autodiscover-cloud/redis-cloud-subscriptions/styles.module.scss
@@ -115,7 +115,7 @@
   font-size: 14px !important;
 
   li {
-    line-height: 20px;
+    line-height: 24px;
   }
 
   .dot {
